update AS settings

+++ android/studio/install.sh
@@ -1,6 +1,13 @@
 #!/usr/bin/env bash
 
+if [ -z $1 ]; then
+	echo "You must specify Android Studio Version: install.sh 3.6"
+	exit 1
+fi
+
 echo "Installing Android Studio version $1 dotfiles .."
+
+exit 1
 
 STUDIO=~/Library/Preferences/AndroidStudio$1
 for file in *; do rm -r $STUDIO/$file; done
